+++ include/memory.h
@@ -9,7 +9,7 @@
 ========================================
 */
 
-void *memset(void *dst, i8 value, usize size);
+void *memset(void *dst, int value, usize size);
 void *memcpy(void *dst, const void *src, usize size);
 void *memmove(void *dst, const void *src, usize size);
